Support decode slice scanner
